+++ app/src/main/java/com/singlelab/bitcoinbrawler/util/Extensions.kt
@@ -3,7 +3,7 @@ package com.singlelab.bitcoinbrawler.util
 import kotlin.math.pow
 import kotlin.math.roundToInt
 
-fun Double.roundTo(numFractionDigits: Int): Double {
+fun Float.roundTo(numFractionDigits: Int): Float {
     val factor = 10.0.pow(numFractionDigits.toDouble())
-    return (this * factor).roundToInt() / factor
+    return (this * factor).roundToInt() / factor.toFloat()
 }
